Display results to console in table
This makes the console output a lot easier to read and understand.

Splits metrics out into their own module so the configuration can be re-used.
