+++ lib/analyse.js
@@ -1,34 +1,4 @@
-const metrics = {
-  'M1': {
-    description: 'Time to First Paint',
-    indexer: 'first',
-    fn: (line) => {
-      return line.message.message.method === 'Tracing.dataCollected' && line.message.message.params.name === 'Paint';
-    }
-  },
-  'M2': {
-    description: 'DomContentLoaded',
-    indexer: 'first',
-    fn: (line) => {
-      return line.message.message.method === 'Page.domContentEventFired';
-    }
-  },
-  /*'M3': {
-    description: 'Render Settlement',
-    indexer: 'last',
-    fn: (line) => {
-      const paint = line.message.message.method === 'Tracing.dataCollected' && line.message.message.params.name === 'Layout';
-      return paint;// && line.message.message.params.args.data.clip[1] < 800;
-    }
-  },*/
-  'M6': {
-    description: 'Page Load',
-    indexer: 'first',
-    fn: (line) => {
-      return line.message.message.method === 'Page.loadEventFired';
-    }
-  }
-};
+const metrics = require('./metrics');
 
 const indexers = {
   first: (data, test) => {
