+++ lib/index.js
@@ -4,6 +4,7 @@ const runner = require('timeliner');
 const analyse = require('./analyse');
 const chart = require('./fps-chart');
 const reporters = require('./reporters');
+const metrics = require('./metrics');
 
 function analyser (opts) {
   if (!opts.url) {
@@ -45,7 +46,7 @@ function analyser (opts) {
       });
       return report;
     })
-    .then((report) => Promise.all(enabledReporters.map((r) => reporters[r].call(this, report))));
+    .then((report) => Promise.all(enabledReporters.map((r) => reporters[r].call(this, report, metrics))));
 }
 
 module.exports = analyser;
