+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "babar": "0.0.3",
     "bluebird": "^3.4.6",
+    "cli-table": "^0.3.1",
     "lodash.meanby": "^4.10.0",
     "minimist": "^1.2.0",
     "timeliner": "^0.4.0"
